# Improve FromEnumerable performance