```
BenchmarkDotNet v0.14.0, Windows 11 (10.0.26100.3775)
13th Gen Intel Core i7-13700F, 1 CPU, 24 logical and 16 physical cores
.NET SDK 10.0.100-preview.3.25201.16
  [Host]            : .NET 9.0.5 (9.0.525.21509), X64 RyuJIT AVX2
  Default(.NET 9.0) : .NET 9.0.5 (9.0.525.21509), X64 RyuJIT AVX2

Job=Default(.NET 9.0)  PowerPlanMode=00000000-0000-0000-0000-000000000000  Toolchain=.NET 9.0  
IterationCount=3  LaunchCount=1  RunStrategy=Throughput  
WarmupCount=3  

| Method                           | Mean     | Error     | StdDev    | Allocated |
|--------------------------------- |---------:|----------:|----------:|----------:|
| IReadOnlyListEnumeratorFromList  | 5.246 us | 0.0611 us | 0.0033 us |      32 B |
| ListEnumerator                   | 2.749 us | 0.1844 us | 0.0101 us |         - |
| IReadOnlyListFor                 | 3.413 us | 0.0591 us | 0.0032 us |         - |
| ZLinqFromEnumerableFromRange     | 7.897 μs | 0.3942 μs | 0.0216 μs |      40 B |
| ZLinqFromEnumerableFromRange(PR) | 7.147 us | 0.0808 us | 0.0044 us |      40 B |
| ZLinqFromEnumerableFromArray     | 5.323 μs | 0.0304 μs | 0.0017 μs |         - |
| ZLinqFromEnumerableFromArray(PR) | 4.778 us | 0.2563 us | 0.0141 us |         - |
| ZLinqFromEnumerableFromList      | 6.703 μs | 0.0583 μs | 0.0032 μs |         - |
| ZLinqFromEnumerableFromList(PR)  | 6.392 us | 0.5306 us | 0.0291 us |         - |
```
```cs
using ZLinq;

namespace Benchmark;

public class FromEnumerableBench
{
    int[] sourceArray = null!;
    List<int> sourceList = null!;
    IReadOnlyList<int> sourceIReadOnlyList = null!;
    IReadOnlyList<int> sourceIReadOnlyListFromList = null!;

    [GlobalSetup]
    public void Setup()
    {
        sourceArray = Enumerable.Range(1, 10000).ToArray();
        sourceList = Enumerable.Range(1, 10000).ToList();
        sourceIReadOnlyList = sourceArray;
        sourceIReadOnlyListFromList = sourceList;
    }

    [Benchmark]
    public int IReadOnlyListEnumeratorFromList()
    {
        int sum = 0;
        var seq = sourceIReadOnlyList;

        foreach (var item in seq)
        {
            sum += item;
        }

        return sum;
    }

    [Benchmark]
    public int ListEnumerator()
    {
        int sum = 0;
        var list = sourceList;
        foreach (var item in list)
        {
            sum += item;
        }

        return sum;
    }

    [Benchmark]
    public int IReadOnlyListFor()
    {
        int sum = 0;
        var list = sourceIReadOnlyListFromList;
        for (var i = 0; i < list.Count; i++)
        {
            sum += list[i];
        }

        return sum;
    }

    [Benchmark]
    public int ZLinqFromEnumerableFromRange()
    {
        int sum = 0;
        var seq = Enumerable.Range(1, 10000)
            .AsValueEnumerable();

        foreach (var item in seq)
        {
            sum += item;
        }

        return sum;
    }

    [Benchmark]
    public int ZLinqFromEnumerableFromArray()
    {
        int sum = 0;
        var seq = ((IEnumerable<int>)sourceArray)
            .AsValueEnumerable();

        foreach (var item in seq)
        {
            sum += item;
        }

        return sum;
    }

    [Benchmark]
    public int ZLinqFromEnumerableFromList()
    {
        int sum = 0;
        var seq = ((IEnumerable<int>)sourceList)
            .AsValueEnumerable();

        foreach (var item in seq)
        {
            sum += item;
        }

        return sum;
    }
}
```